Fix Tags on items with similar class names
Fix escaping of class name in search of items. It can happen that you can have tags defined for "TicketRecurrent" and they will be shown also for the "Ticket".
Escape with double quotes. Question is whether it can happen it will be stored in DB also without "" (so not the JSON).
At least I can see in DB it is stored also for single type tags like ["Ticket"] so fix is working for me.